# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BRAX Marketplace API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cart/items:
    post:
      operationId: addCartItem
      tags: [cart]
      summary: Add item to cart (with live inventory check)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCartItemBody"
      responses:
        "201":
          description: Item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "409":
          description: Insufficient inventory
  /cart/items/{cartItemId}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - in: path
          name: cartItemId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - in: path
          name: cartItemId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  # ── ORDERS ──────────────────────────────────────────────────────────────────
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List user orders
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        - in: query
          name: page
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedOrders"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order (checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}/status:
    put:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (seller/admin)
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ── PAYMENTS ────────────────────────────────────────────────────────────────
  /payments/stripe/create-intent:
    post:
      operationId: createStripePaymentIntent
      tags: [payments]
      summary: Create a Stripe payment intent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStripeIntentBody"
      responses:
        "200":
          description: Payment intent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StripePaymentIntent"
  /payments/flutterwave/initiate:
    post:
      operationId: initiateFlutterwavePayment
      tags: [payments]
      summary: Initiate Flutterwave mobile money payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiateFlutterwaveBody"
      responses:
        "200":
          description: Payment link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlutterwavePaymentLink"
...
components:
  schemas:
...
    Cart:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        itemCount:
          type: integer
      required: [id, userId, items, subtotal, itemCount]
    CartItem:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        productName:
          type: string
        productImageUrl:
          type: string
        price:
          type: number
        quantity:
          type: integer
        availableStock:
          type: integer
      required: [id, productId, productName, productImageUrl, price, quantity, availableStock]
    AddCartItemBody:
      type: object
      properties:
        productId:
          type: string
        quantity:
          type: integer
      required: [productId, quantity]
    UpdateCartItemBody:
      type: object
      properties:
        quantity:
          type: integer
      required: [quantity]
    Order:
      type: object
      properties:
        id:
          type: string
        buyerId:
          type: string
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        paymentStatus:
          type: string
          enum: [pending, paid, failed, refunded]
        paymentMethod:
          type: string
          enum: [stripe, flutterwave, pending]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        shippingCost:
          type: number
        total:
          type: number
        shippingAddress:
          $ref: "#/components/schemas/Address"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, buyerId, status, paymentStatus, paymentMethod, items, subtotal, shippingCost, total, shippingAddress, createdAt, updatedAt]
    OrderItem:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        productName:
          type: string
        productImageUrl:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
      required: [id, productId, productName, productImageUrl, quantity, unitPrice, totalPrice]
    PaginatedOrders:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [items, total, page, totalPages]
    CreateOrderBody:
      type: object
      properties:
        shippingAddress:
          $ref: "#/components/schemas/Address"
        paymentMethod:
          type: string
          enum: [stripe, flutterwave]
        cartId:
          type: string
      required: [shippingAddress, paymentMethod, cartId]
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [confirmed, processing, shipped, delivered, cancelled]
      required: [status]
    Address:
      type: object
      properties:
        fullName:
          type: string
        line1:
          type: string
        line2:
          type: string
          nullable: true
        city:
          type: string
        state:
          type: string
        country:
          type: string
        postalCode:
          type: string
        phone:
          type: string
      required: [fullName, line1, city, state, country, postalCode, phone]
...
    CreateStripeIntentBody:
      type: object
      properties:
        orderId:
          type: string
        amount:
          type: number
        currency:
          type: string
          default: "usd"
      required: [orderId, amount]
    StripePaymentIntent:
      type: object
      properties:
        clientSecret:
          type: string
        paymentIntentId:
          type: string
      required: [clientSecret, paymentIntentId]
    InitiateFlutterwaveBody:
      type: object
      properties:
        orderId:
          type: string
        amount:
          type: number
        currency:
          type: string
          default: "NGN"
        email:
          type: string
        phone:
          type: string
        name:
          type: string
      required: [orderId, amount, currency, email, phone, name]
    FlutterwavePaymentLink:
      type: object
      properties:
        paymentLink:
          type: string
        transactionRef:
          type: string
      required: [paymentLink, transactionRef]